requirements-dev.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
dj-static==0.0.6
Django==4.0.6
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
unicodecsv==0.14.1
urllib3==1.26.11
whitenoise==6.2.0
